requirements.txt
dj-database-url==0.5.0
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==5.0.1

